backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
elasticsearch==8.15.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
